feat(648): extract add endpoint modal into separate component
